Unit test file and dependencies, unfinished
SCAUnitTests.cpp has unit tests for each of the defined functions currently in SCA.cpp. Other files are dependencies, pch.cpp and pch.h for functionality and testMain.cpp and testNoMain.cpp for test files to run against.

Currently, some of the tests aren't where they need to be. I need a sample template file for matchTemplateFileWithTree to be tested and to know what it's meant to return, and I need to get a better understanding of readANTLROutputTree to write a test for it still.