Adding more test cpp files
Surprisingly little amount of cpp files from my previous classes that were useful. bankSystem was my final project for CSIT121 and is very long, surely full of little errors. studentFiles was from the same class but earlier on. testFile is a file I made up just now for the purpose of testing, but it looks like an entry level program with user defined functions.